Refactor apierrors package and AddDimension func to dimension.Add
