+++ dimension/dimension_test.go
@@ -7,7 +7,7 @@ import (
 	"net/http/httptest"
 	"testing"
 
-	"github.com/ONSdigital/dp-dataset-api/api-errors"
+	errs "github.com/ONSdigital/dp-dataset-api/apierrors"
 	"github.com/ONSdigital/dp-dataset-api/dimension"
 	"github.com/ONSdigital/dp-dataset-api/models"
 	"github.com/ONSdigital/dp-dataset-api/store/datastoretest"
@@ -18,17 +18,16 @@ const secretKey = "coffee"
 
 var internalError = errors.New("internal error")
 
-func createRequestWithToken(method, url string, body io.Reader) (*http.Request, error) {
-	r, err := http.NewRequest(method, url, body)
+func createRequestWithToken(method, url string, body io.Reader) *http.Request {
+	r := httptest.NewRequest(method, url, body)
 	r.Header.Add("internal-token", secretKey)
-	return r, err
+	return r
 }
 
 func TestAddNodeIDToDimensionReturnsOK(t *testing.T) {
 	t.Parallel()
 	Convey("Add node id to a dimension returns ok", t, func() {
-		r, err := createRequestWithToken("PUT", "http://localhost:21800/instances/123/dimensions/age/options/55/node_id/11", nil)
-		So(err, ShouldBeNil)
+		r := createRequestWithToken("PUT", "http://localhost:21800/instances/123/dimensions/age/options/55/node_id/11", nil)
 		w := httptest.NewRecorder()
 
 		mockedDataStore := &storetest.StorerMock{
@@ -48,13 +47,12 @@ func TestAddNodeIDToDimensionReturnsOK(t *testing.T) {
 func TestAddNodeIDToDimensionReturnsBadRequest(t *testing.T) {
 	t.Parallel()
 	Convey("Add node id to a dimension returns bad request", t, func() {
-		r, err := createRequestWithToken("PUT", "http://localhost:21800/instances/123/dimensions/age/options/55/node_id/11", nil)
-		So(err, ShouldBeNil)
+		r := createRequestWithToken("PUT", "http://localhost:21800/instances/123/dimensions/age/options/55/node_id/11", nil)
 		w := httptest.NewRecorder()
 
 		mockedDataStore := &storetest.StorerMock{
 			UpdateDimensionNodeIDFunc: func(event *models.Dimension) error {
-				return api_errors.DimensionNodeNotFound
+				return errs.DimensionNodeNotFound
 			},
 		}
 
@@ -69,8 +67,7 @@ func TestAddNodeIDToDimensionReturnsBadRequest(t *testing.T) {
 func TestAddNodeIDToDimensionReturnsInternalError(t *testing.T) {
 	t.Parallel()
 	Convey("Add node id to a dimension returns internal error", t, func() {
-		r, err := createRequestWithToken("PUT", "http://localhost:21800/instances/123/dimensions/age/options/55/node_id/11", nil)
-		So(err, ShouldBeNil)
+		r := createRequestWithToken("PUT", "http://localhost:21800/instances/123/dimensions/age/options/55/node_id/11", nil)
 		w := httptest.NewRecorder()
 
 		mockedDataStore := &storetest.StorerMock{
@@ -87,11 +84,70 @@ func TestAddNodeIDToDimensionReturnsInternalError(t *testing.T) {
 	})
 }
 
+func TestAddDimensionToInstanceReturnsOk(t *testing.T) {
+	t.Parallel()
+	Convey("Add a dimension to an instance returns ok", t, func() {
+		r := createRequestWithToken("PUT", "http://localhost:21800/instances/123/dimensions/age/options/55", nil)
+		w := httptest.NewRecorder()
+
+		mockedDataStore := &storetest.StorerMock{
+			AddDimensionToInstanceFunc: func(event *models.Dimension) error {
+				return nil
+			},
+		}
+
+		dimension := &dimension.Store{mockedDataStore}
+		dimension.Add(w, r)
+
+		So(w.Code, ShouldEqual, http.StatusOK)
+		So(len(mockedDataStore.AddDimensionToInstanceCalls()), ShouldEqual, 1)
+	})
+}
+
+func TestAddDimensionToInstanceReturnsNotFound(t *testing.T) {
+	t.Parallel()
+	Convey("Add a dimension to an instance returns not found", t, func() {
+		r := createRequestWithToken("PUT", "http://localhost:21800/instances/123/dimensions/age/options/55", nil)
+		w := httptest.NewRecorder()
+
+		mockedDataStore := &storetest.StorerMock{
+			AddDimensionToInstanceFunc: func(event *models.Dimension) error {
+				return errs.DimensionNodeNotFound
+			},
+		}
+
+		dimension := &dimension.Store{mockedDataStore}
+		dimension.Add(w, r)
+
+		So(w.Code, ShouldEqual, http.StatusNotFound)
+		So(len(mockedDataStore.AddDimensionToInstanceCalls()), ShouldEqual, 1)
+	})
+}
+
+func TestAddDimensionToInstanceReturnsInternalError(t *testing.T) {
+	t.Parallel()
+	Convey("Add a dimension to an instance returns internal error", t, func() {
+		r := createRequestWithToken("PUT", "http://localhost:21800/instances/123/dimensions/age/options/55", nil)
+		w := httptest.NewRecorder()
+
+		mockedDataStore := &storetest.StorerMock{
+			AddDimensionToInstanceFunc: func(event *models.Dimension) error {
+				return internalError
+			},
+		}
+
+		dimension := &dimension.Store{mockedDataStore}
+		dimension.Add(w, r)
+
+		So(w.Code, ShouldEqual, http.StatusInternalServerError)
+		So(len(mockedDataStore.AddDimensionToInstanceCalls()), ShouldEqual, 1)
+	})
+}
+
 func TestGetDimensionNodesReturnsOk(t *testing.T) {
 	t.Parallel()
 	Convey("Get dimension nodes returns ok", t, func() {
-		r, err := createRequestWithToken("GET", "http://localhost:21800/instances/123/dimensions", nil)
-		So(err, ShouldBeNil)
+		r := createRequestWithToken("GET", "http://localhost:21800/instances/123/dimensions", nil)
 		w := httptest.NewRecorder()
 
 		mockedDataStore := &storetest.StorerMock{
@@ -111,13 +167,12 @@ func TestGetDimensionNodesReturnsOk(t *testing.T) {
 func TestGetDimensionNodesReturnsNotFound(t *testing.T) {
 	t.Parallel()
 	Convey("Get dimension nodes returns not found", t, func() {
-		r, err := createRequestWithToken("GET", "http://localhost:21800/instances/123/dimensions", nil)
-		So(err, ShouldBeNil)
+		r := createRequestWithToken("GET", "http://localhost:21800/instances/123/dimensions", nil)
 		w := httptest.NewRecorder()
 
 		mockedDataStore := &storetest.StorerMock{
 			GetDimensionNodesFromInstanceFunc: func(id string) (*models.DimensionNodeResults, error) {
-				return nil, api_errors.InstanceNotFound
+				return nil, errs.InstanceNotFound
 			},
 		}
 
@@ -132,8 +187,7 @@ func TestGetDimensionNodesReturnsNotFound(t *testing.T) {
 func TestGetDimensionNodesReturnsInternalError(t *testing.T) {
 	t.Parallel()
 	Convey("Get dimension nodes returns internal error", t, func() {
-		r, err := createRequestWithToken("GET", "http://localhost:21800/instances/123/dimensions", nil)
-		So(err, ShouldBeNil)
+		r := createRequestWithToken("GET", "http://localhost:21800/instances/123/dimensions", nil)
 		w := httptest.NewRecorder()
 
 		mockedDataStore := &storetest.StorerMock{
@@ -153,8 +207,7 @@ func TestGetDimensionNodesReturnsInternalError(t *testing.T) {
 func TestGetUniqueDimensionValuesReturnsOk(t *testing.T) {
 	t.Parallel()
 	Convey("Get all unique dimensions returns ok", t, func() {
-		r, err := createRequestWithToken("GET", "http://localhost:21800/instances/123/dimensions/age/options", nil)
-		So(err, ShouldBeNil)
+		r := createRequestWithToken("GET", "http://localhost:21800/instances/123/dimensions/age/options", nil)
 		w := httptest.NewRecorder()
 
 		mockedDataStore := &storetest.StorerMock{
@@ -179,7 +232,7 @@ func TestGetUniqueDimensionValuesReturnsNotFound(t *testing.T) {
 		w := httptest.NewRecorder()
 		mockedDataStore := &storetest.StorerMock{
 			GetUniqueDimensionValuesFunc: func(id, dimension string) (*models.DimensionValues, error) {
-				return nil, api_errors.InstanceNotFound
+				return nil, errs.InstanceNotFound
 			},
 		}
